Refactor backend code (#509)
* separate spire and tornjak APIs


* refactored api/agent/server.go


* Fix CodeQL check


* nit indentation


---------
